Revert "removed login requirement"
This reverts commit eee92a52f05c7d0a01833cfced14c78157f74ddf.

+++ app.py
@@ -45,7 +45,7 @@ def showCategories():
 
 # Category CRUD operations
 @app.route('/catalog/new', methods=['POST','GET'])
-# @login_required
+@login_required
 def newCategories():
     categories = session.query(Category).order_by(asc(Category.name))
     if request.method == 'POST':
@@ -58,13 +58,13 @@ def newCategories():
         return render_template('newCategory.html',categories=categories)
 
 @app.route('/catalog/<int:category_id>/edit/', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def editCategories(category_id):
     categories = session.query(Category).order_by(asc(Category.name))
     editCategories = session.query(Category).filter_by(id=category_id).one()
-    # if editCategories != login_session['user_id']:
-    #     flash(u'You are not authorized to edit this category', 'error')
-    #     return redirect(url_for('showCategories'))
+    if editCategories != login_session['user_id']:
+        flash(u'You are not authorized to edit this category', 'error')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         if request.form['name']:
             editCategories.name = request.form['name']
@@ -78,13 +78,13 @@ def editCategories(category_id):
                                 category=editCategories)
 
 @app.route('/catalog/<int:category_id>/delete/', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def deleteCategories(category_id):
     categories = session.query(Category).order_by(asc(Category.name))
     deleteCategory = session.query(Category).filter_by(id=category_id).one()
-    # if deleteCategories != login_session['user_id']:
-    #     flash(u'You are not authorized to delete this category', 'error')
-    #     return redirect(url_for('showCategories'))
+    if deleteCategories != login_session['user_id']:
+        flash(u'You are not authorized to delete this category', 'error')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         session.delete(deleteCategory)
         flash('\"%s\" Successfully Deleted' % deleteCategory.name)
@@ -116,7 +116,7 @@ def itemPage(category_id,item_id):
                             item_id=item_id,
                             categories=categories)
 @app.route('/catalog/<int:category_id>/new', methods=['POST','GET'])
-# @login_required
+@login_required
 def newItem(category_id):
     category = session.query(Category).filter_by(id=category_id).one()    
     if request.method == 'POST':
@@ -135,14 +135,14 @@ def newItem(category_id):
                                 category=category)
 
 @app.route('/catalog/<int:category_id>/items/<int:item_id>/edit', methods=['POST','GET'])
-# @login_required
+@login_required
 def editItem(category_id,item_id):
     categories = session.query(Category).order_by(asc(Category.name))
     category = session.query(Category).filter_by(id=category_id).one()
     editItem = session.query(Item).filter_by(id=item_id).one()
-    # if editItem != login_session['user_id']:
-    #     flash(u'You are not authorized to edit this item', 'error')
-    #     return redirect(url_for('showCategories'))
+    if editItem != login_session['user_id']:
+        flash(u'You are not authorized to edit this item', 'error')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         if request.form['name']:
             editItem.name = request.form['name']
@@ -164,12 +164,12 @@ def editItem(category_id,item_id):
                                 categories=categories)
 
 @app.route('/catalog/<int:category_id>/items/<int:item_id>/delete', methods=['POST','GET'])
-# @login_required
+@login_required
 def deleteItem(category_id,item_id):
     deleteItem = session.query(Item).filter_by(id=item_id).one()
-    # if deleteItem != login_session['user_id']:
-    #     flash(u'You are not authorized to delete this item', 'error')
-    #     return redirect(url_for('showCategories'))
+    if deleteItem != login_session['user_id']:
+        flash(u'You are not authorized to delete this item', 'error')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         session.delete(deleteItem)
         flash('\"%s\" Successfully Deleted' % deleteItem.name)
@@ -206,131 +206,131 @@ CLIENT_ID = json.loads(
     open('fb_client_secrets.json', 'r').read())['web']['app_id']
 APPLICATION_NAME = "Item Catalog Application"
 
-# @app.route('/login/')
-# def showLogin():
-#     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
-#                     for x in xrange(32))
-#     login_session['state'] = state
-#     # return "The current session state is %s" % login_session['state']
-#     return render_template('login.html', STATE=state)
+@app.route('/login/')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    # return "The current session state is %s" % login_session['state']
+    return render_template('login.html', STATE=state)
 
-# @app.route('/fbconnect', methods=['POST'])
-# def fbconnect():
-#     if request.args.get('state') != login_session['state']:
-#         response = make_response(json.dumps('Invalid state parameter.'), 401)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-#     access_token = request.data
-#     print "access token received %s " % access_token
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s " % access_token
 
-#     app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
-#         'web']['app_id']
-#     app_secret = json.loads(
-#         open('fb_client_secrets.json', 'r').read())['web']['app_secret']
-#     url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
-#         app_id, app_secret, access_token)
-#     h = httplib2.Http()
-#     result = h.request(url, 'GET')[1]
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads(
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
 
-#     # Use token to get user info from API
-#     userinfo_url = "https://graph.facebook.com/v2.4/me"
-#     # strip expire tag from access token
-#     token = result.split("&")[0]
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.4/me"
+    # strip expire tag from access token
+    token = result.split("&")[0]
 
 
-    # url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
-    # h = httplib2.Http()
-    # result = h.request(url, 'GET')[1]
-    # # print "url sent for API access:%s"% url
-    # # print "API JSON result: %s" % result
-    # data = json.loads(result)
-    # login_session['provider'] = 'facebook'
-    # login_session['username'] = data["name"]
-    # login_session['email'] = data["email"]
-    # login_session['facebook_id'] = data["id"]
+    url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
 
-    # # The token must be stored in the login_session in order to properly logout, let's strip out the information before the equals sign in our token
-    # stored_token = token.split("=")[1]
-    # login_session['access_token'] = stored_token
+    # The token must be stored in the login_session in order to properly logout, let's strip out the information before the equals sign in our token
+    stored_token = token.split("=")[1]
+    login_session['access_token'] = stored_token
 
-    # # Get user picture
-    # url = 'https://graph.facebook.com/v2.4/me/picture?%s&redirect=0&height=200&width=200' % token
-    # h = httplib2.Http()
-    # result = h.request(url, 'GET')[1]
-    # data = json.loads(result)
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.4/me/picture?%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
 
-    # login_session['picture'] = data["data"]["url"]
+    login_session['picture'] = data["data"]["url"]
 
-    # # see if user exists
-    # user_id = getUserID(login_session['email'])
-    # if not user_id:
-    #     user_id = createUser(login_session)
-    # login_session['user_id'] = user_id
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
 
-    # output = ''
-    # output += '<h1>Welcome, '
-    # output += login_session['username']
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
 
-    # output += '!</h1>'
-    # output += '<img src="'
-    # output += login_session['picture']
-    # output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
 
-    # flash("Now logged in as %s" % login_session['username'])
-    # return output
+    flash("Now logged in as %s" % login_session['username'])
+    return output
 
 # DISCONNECT - Revoke a current user's token and reset their login_session
 
-# @app.route('/fbdisconnect')
-# def fbdisconnect():
-#     facebook_id = login_session['facebook_id']
-#     # The access token must me included to successfully logout
-#     access_token = login_session['access_token']
-#     url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
-#     h = httplib2.Http()
-#     result = h.request(url, 'DELETE')[1]
-#     return "you have been logged out"
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "you have been logged out"
 
 
-# @app.route('/disconnect')
-# def disconnect():
-#     if 'provider' in login_session:
-#         if login_session['provider'] == 'facebook':
-#             fbdisconnect()
-#             del login_session['facebook_id']
-#         del login_session['username']
-#         del login_session['email']
-#         del login_session['picture']
-#         del login_session['user_id']
-#         del login_session['provider']
-#         flash("You have successfully been logged out.")
-#         return redirect(url_for('showCategories'))
-#     else:
-#         flash("You were not logged in")
-#         return redirect(url_for('showCategories'))
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully been logged out.")
+        return redirect(url_for('showCategories'))
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('showCategories'))
 # User Helper Functions
 
-# def createUser(login_session):
-#     newUser = User(name=login_session['username'], 
-#                    email=login_session['email'], 
-#                    picture=login_session['picture'])
-#     session.add(newUser)
-#     session.commit()
-#     user = session.query(User).filter_by(email=login_session['email']).one()
-#     return user.id
+def createUser(login_session):
+    newUser = User(name=login_session['username'], 
+                   email=login_session['email'], 
+                   picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
 
 
-# def getUserInfo(user_id):
-#     user = session.query(User).filter_by(id=user_id).one()
-#     return user
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
 
 
-# def getUserID(email):
-#     try:
-#         user = session.query(User).filter_by(email=email).one()
-#         return user.id
-#     except:
-#         return None
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
 
 
 if __name__ == '__main__':
